chore: replace com.example with com.hahaha
