+++ boot-receive/src/main/java/com/hahaha/SendApp.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.hahaha;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
