+++ boot-receive/src/main/java/com/hahaha/listener/MyRabbitListener.java
@@ -1,4 +1,4 @@
-package com.example.listener;
+package com.hahaha.listener;
 
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
