+++ boot-send/src/test/java/com/hahaha/config/MyTest.java
@@ -1,4 +1,5 @@
-package com.example.config;
+package com.hahaha.config;
+
 
 
 import org.junit.jupiter.api.Test;
@@ -24,14 +25,9 @@ class MyTest {
     private AmqpTemplate amqpTemplate;
 
     @Test
-    void receive() {
-        /**
-         * receiveAndConvert:主动获取队列中的消息
-         * 参数:队列名字
-         * 返回值:返回队列中最前面的一条消息,如果队列中没有消息,则返回空
-         */
-        String message = (String) amqpTemplate.receiveAndConvert("bootQueue");
-        System.out.println(message);
+    void send() {
+        String message = "boot消息001";
+        amqpTemplate.convertAndSend("bootExchange","bootKey", message);
     }
 
 }
