+++ java-receive/src/main/java/com/hahaha/DirectReceive.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.hahaha;
 
 import com.rabbitmq.client.*;
