+++ java-send/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>cn.wolfcode.wolf2w</groupId>
+    <groupId>com.hahaha</groupId>
     <artifactId>java-send</artifactId>
     <version>1.0-SNAPSHOT</version>
