+++ java-send/src/main/java/com/hahaha/DirectSend.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.hahaha;
 
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
@@ -8,7 +8,7 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.TimeoutException;
 
-public class FanoutSend {
+public class DirectSend {
     public static void main(String[] args) {
         ConnectionFactory factory = new ConnectionFactory();
         factory.setHost("192.168.171.130");
@@ -25,7 +25,7 @@ public class FanoutSend {
             //在物理链接中创建虚拟信道
             channel = connection.createChannel();
 
-            String queueName = "fanoutQueue";
+            String queueName = "directQueue";
             /**
              * 定义队列,如果队列不存在创建,存在就直接调用
              * 参数1:队列名
@@ -36,14 +36,14 @@ public class FanoutSend {
              */
             channel.queueDeclare(queueName, true, false, false, null);
 
-            String exchangeName = "fanoutExchange";
+            String exchangeName = "directExchange";
             /**
              * 交换机
              * 参数1:交换机名字
              * 参数2:交换机类型(direct topic fanout)
              * 参数3:是否持久化
              */
-            channel.exchangeDeclare(exchangeName, "fanout", true);
+            channel.exchangeDeclare(exchangeName, "direct", true);
 
             /**
              * 绑定队列
@@ -51,9 +51,9 @@ public class FanoutSend {
              * 参数2: 交换机名
              * 参数3: 路由键
              */
-            channel.queueBind(queueName, exchangeName, "fanout");
+            channel.queueBind(queueName, exchangeName, "directKey");
 
-            String message = "hello fanout exchange 消息002";
+            String message = "hello direct exchange 消息002";
             /**
              * 发送消息
              * 参数1: 交换机名字
@@ -61,7 +61,7 @@ public class FanoutSend {
              * 参数3: 属性,null即可
              * 参数4: 消息内容
              */
-            channel.basicPublish(exchangeName, "fanoutKey", null,
+            channel.basicPublish(exchangeName, "directKey", null,
                     message.getBytes(StandardCharsets.UTF_8));
 
             System.out.println("消息已发送：" + message);
