+++ java-send/src/main/java/com/hahaha/TopicSend.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.hahaha;
 
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
