+++ java-send/src/main/java/com/hahaha/WaitForConfirmSend.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.hahaha;
 
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
@@ -8,7 +8,7 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.TimeoutException;
 
-public class TxSend {
+public class WaitForConfirmSend {
     public static void main(String[] args) {
         ConnectionFactory factory = new ConnectionFactory();
         factory.setHost("192.168.171.130");
@@ -20,28 +20,28 @@ public class TxSend {
         Connection connection = null;
         Channel channel = null;
         try {
-            // 创建物理连接
+            //创建物理连接
             connection = factory.newConnection();
-            // 在物理连接中创建虚拟信道
+            //在物理链接中创建虚拟信道
             channel = connection.createChannel();
 
             String queueName = "directQueue";
             /**
-             * 定义队列, 如果队列不存在创建, 存在就直接调用
-             * 参数1: 队列名
-             * 参数2: 是否持久化, true 表示持久化
-             * 参数3: 是否排他
-             * 参数4: 无消费者监听该队列, 是否自动删除
-             * 参数5: 属性, 一般填 null 即可
+             * 定义队列,如果队列不存在创建,存在就直接调用
+             * 参数1:队列名
+             * 参数2: 是否持久化,true表示持久化
+             * 参数3:是否排他
+             * 参数4:无消费者监听该队列,是否自动删除
+             * 参数5:属性,一般填null即可
              */
             channel.queueDeclare(queueName, true, false, false, null);
 
             String exchangeName = "directExchange";
             /**
              * 交换机
-             * 参数1: 交换机名字
-             * 参数2: 交换机类型 (direct, topic, fanout)
-             * 参数3: 是否持久化
+             * 参数1:交换机名字
+             * 参数2:交换机类型(direct topic fanout)
+             * 参数3:是否持久化
              */
             channel.exchangeDeclare(exchangeName, "direct", true);
 
@@ -52,51 +52,44 @@ public class TxSend {
              * 参数3: 路由键
              */
             channel.queueBind(queueName, exchangeName, "directKey");
-
-            // 开启事务
-            channel.txSelect();
-
+//开启发送者确认模式
+            channel.confirmSelect();
             String message = "hello direct exchange 消息001";
             /**
              * 发送消息
              * 参数1: 交换机名字
              * 参数2: 路由键
-             * 参数3: 属性, null 即可
+             * 参数3: 属性,null即可
              * 参数4: 消息内容
              */
-            channel.basicPublish(
-                    exchangeName,
-                    "directKey",
-                    null,
-                    message.getBytes(StandardCharsets.UTF_8)
-            );
-
-            // 提交事务
-            channel.txCommit();
+            channel.basicPublish(exchangeName, "directKey", null,
+                    message.getBytes(StandardCharsets.UTF_8));
+            try {
+                boolean b = channel.waitForConfirms(1000l);
+                if (!b) {
+                    System.out.println("补发消息");
+                }
+            } catch (InterruptedException e) {
+                System.out.println("不能确定消息是否发送成功,需要补发消息,但是可能造成重复消息,需要在消费者代码中实现操作的幂等性");
+            } catch (TimeoutException e) {
+                System.out.println("不能确定消息是否发送成功,需要补发消息,但是可能造成重复消息,需要在消费者代码中实现操作的幂等性");
+            }
             System.out.println("消息已发送：" + message);
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (TimeoutException e) {
+        } catch (IOException | TimeoutException e) {
             e.printStackTrace();
         } finally {
-            if (channel != null) {
+            // 先关闭信道
+            if (channel != null && channel.isOpen()) {
                 try {
-                    // 回滚事务并关闭信道
-                    channel.txRollback();
                     channel.close();
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                } catch (TimeoutException e) {
-                    throw new RuntimeException(e);
+                } catch (Exception ignore) {
                 }
             }
-            if (connection != null) {
+            // 再关闭连接
+            if (connection != null && connection.isOpen()) {
                 try {
-                    // 关闭连接
                     connection.close();
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
+                } catch (Exception ignore) {
                 }
             }
         }
